rework to use cmd structure, worker now runs as seperate service as it would in kube context
