+++ Dockerfile
@@ -5,6 +5,6 @@ WORKDIR /app/
 
 RUN apk add git
 RUN go mod tidy
-RUN chmod 777 init.sh
+RUN chmod 777 start.sh
 
 EXPOSE 1323
