--- init.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-go build .
-./visibilityreport
