+++ start.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+if [ $1 = "worker" ]; then
+  cd ./cmd/worker
+  go build .
+  ./worker
+else
+  cd ./cmd/api
+  go build .
+  ./api
+fi
